refactor: move data related functions to dataset.py
